requirements.txt
python-telegram-bot==20.6
functions-framework==3.*
google-cloud-firestore==2.12.0